style: update animation and padding in CSS
The animation for the drop-in effect has been updated to start with a width of 0px and gradually increase to 30rem, while also adding padding of 1rem. 

The padding has also been adjusted for the popover-inner class, removing the padding on the outer element and adding it to the inner element. Additionally, the width has been set to 0px.